+++ tests/conftest.py
@@ -3,6 +3,7 @@ import os
 import pytest
 import runez
 from runez.conftest import cli, logged, temp_folder
+from runez.http import GlobalHttpCalls
 from runez.pyenv import PythonDepot
 
 from pickley import DOT_META, PICKLEY, PickleyConfig
@@ -11,10 +12,12 @@ from pickley.cli import main
 
 cli.default_exe = PickleyConfig.program_path
 cli.default_main = main
+GlobalHttpCalls.forbid()
 PythonDepot.use_path = False
 assert logged, temp_folder  # Just making fixtures available, with no complaints about unused imports
 
 PickleyConfig.program_path = PickleyConfig.program_path.replace("pytest", PICKLEY)
+PickleyConfig._pickley_dev_path = False
 
 
 def dot_meta(relative=None, parent=None):
@@ -26,14 +29,6 @@ def dot_meta(relative=None, parent=None):
         path = os.path.join(parent, path)
 
     return path
-
-
-def verify_abort(func, *args, **kwargs):
-    exception = kwargs.pop("exception", SystemExit)
-    with runez.CaptureOutput() as logged:
-        with pytest.raises(exception):
-            func(*args, **kwargs)
-        return str(logged)
 
 
 class TemporaryBase(runez.TempFolder):