+++ tests/test_delivery.py
@@ -4,8 +4,9 @@ import pytest
 import runez
 from mock import MagicMock, patch
 
-from pickley import PackageSpec, PickleyConfig
+from pickley import __version__, PackageSpec, PICKLEY, PickleyConfig
 from pickley.delivery import auto_uninstall, DeliveryMethod
+from pickley.package import PythonVenv
 
 
 BREW_INSTALL = "/brew/install/bin"
@@ -71,3 +72,42 @@ def test_uninstall_brew(*_):
         with pytest.raises(SystemExit):
             auto_uninstall("%s/wget" % BREW_INSTALL)
         assert "Can't automatically uninstall" in logged.pop()
+
+
+class SimulatedInstallation:
+    def __init__(self, cfg, name, version):
+        self.cfg = cfg
+        self.name = name
+        self.version = version
+        self.entry_points = {name: name}
+        self.pspec = PackageSpec(cfg, name, version=version)
+        self.pspec.save_manifest(self.entry_points)
+        self.venv = PythonVenv(self.pspec, create=False)
+        self.venv_exe = os.path.join(self.pspec.install_path, "bin", name)
+        runez.write(self.venv_exe, "#!/bin/bash\necho %s" % version)
+        runez.make_executable(self.venv_exe)
+
+    def check_wrap(self, wrap_method):
+        impl = DeliveryMethod.delivery_method_by_name(wrap_method)
+        impl.install(self.pspec, self.venv, self.entry_points)
+        exe = runez.resolved_path(self.name)
+        r = runez.run(exe, "--version", fatal=False)
+        assert r.succeeded
+        assert r.output == self.version
+
+    def check_alternating(self, logged):
+        self.check_wrap("wrap")
+        assert "Wrapped %s -> " % self.name in logged.pop()
+        self.check_wrap("symlink")
+        assert "Symlinked %s -> " % self.name in logged.pop()
+        self.check_wrap("wrap")
+        assert "Wrapped %s -> " % self.name in logged.pop()
+
+
+def test_wrapper(temp_cfg, logged):
+    """Check that flip-flopping between symlink/wrapper works"""
+    mgit = SimulatedInstallation(temp_cfg, "mgit", "1.0")
+    mgit.check_alternating(logged)
+
+    pickley = SimulatedInstallation(temp_cfg, PICKLEY, __version__)
+    pickley.check_alternating(logged)